Change evidence note compliance year to be int to match datetime.year

+++ src/EA.Weee.Core/AatfEvidence/TransferEvidenceNoteData.cs
@@ -21,7 +21,7 @@
 
         public DateTime? ApprovedDate { get; set; }
 
-        public short ComplianceYear { get; set; }
+        public int ComplianceYear { get; set; }
 
         public List<TransferEvidenceNoteTonnageData> TransferEvidenceNoteTonnageData { get; set; }
